+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.0]
+
+### Added
+
+- adaptive height of BaseTextarea
+- maximum width of 100% BaseTextarea during resizing
+
+### Fixed
+
+- bug opening of BaseSelect in disabled mode
+
 ## [0.9.0]
 
 ### Changed